categoria dao

+++ Food Store/src/dao/DAO.java
@@ -7,8 +7,8 @@ import java.util.List;
 public interface DAO <T> {
     List<T> listar();
     void crear(T entidad);
-    void editar(int id);
-    void eliminar(int id);
-    T buscarPorId(int id);
+    void editar(long id);
+    void eliminar(long id);
+    T buscarPorId(long id);
     void actualizar(T entidad);
 }
